requirements.txt
Django==5.1.2
python-telegram-bot==20.7
dj-database-url
psycopg2-binary
gunicorn
whitenoise
requests
httpx>=0.24.0
python-dotenv
nest-asyncio==1.5.8
pytz
apscheduler  